requirements.txt
Flask==3.0.0
graphene>=3.1
flask_sqlalchemy>=3.1
python-dotenv==1.0.1
psycopg2==2.9.9